requirements.txt
streamlit==1.35.0
numpy
opencv-python-headless
Pillow
tensorflow==2.11.0
matplotlib
